Create README.md
A blog application with user login and logout for authors and admins. 
A blog post is a car added by its owner(author).
Posts contain text and an uploaded image by an author. 
Only author or admins can update or delete a post. 
Car list page has a filtered search where an user can choose a category.
Each post in car list page is a link to its own page where it's presented on full screen with more details about the car, and all the coments.

This is just a simple project, created to use my knowledge of python/django with a little bit of styling in CSS. 
I plan to continue working on this project in future and keep improving it untill its ready for deployment.